Enh: relocate spectral estimation code to analysis
